only works with jwtUtil.validateToken(authHeader);
