--- gateway/src/main/java/com/fedkoroma/gateway/enums/Roles.java
@@ -1,11 +1,0 @@
-package com.fedkoroma.gateway.enums;
-
-public enum Roles {
-    USER("USER"),ADMIN("ADMIN");
-
-    public final String label;
-
-    private Roles(String label) {
-        this.label = label;
-    }
-}
