--- gateway/src/main/java/com/fedkoroma/gateway/enums/Urls.java
@@ -1,12 +1,0 @@
-package com.fedkoroma.gateway.enums;
-
-public enum Urls {
-    GET_DATA("main");
-
-    public final String label;
-
-    private Urls(String label) {
-        this.label = label;
-    }
-
-}
