+++ gateway/src/main/java/com/fedkoroma/gateway/filter/AuthenticationFilter.java
@@ -1,72 +1,57 @@
 package com.fedkoroma.gateway.filter;
 
-//import com.fedkoroma.gateway.util.JwtUtil;
-//import io.jsonwebtoken.Claims;
-//import org.apache.commons.lang.StringUtils;
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.cloud.gateway.filter.GatewayFilter;
-//import org.springframework.cloud.gateway.filter.GatewayFilterChain;
-//import org.springframework.http.HttpStatus;
-//import org.springframework.http.server.reactive.ServerHttpRequest;
-//import org.springframework.http.server.reactive.ServerHttpResponse;
-//import org.springframework.stereotype.Component;
-//import org.springframework.web.server.ServerWebExchange;
-//import reactor.core.publisher.Mono;
-//
-//@Component
-//public class AuthenticationFilter implements GatewayFilter {
-//
+import com.fedkoroma.gateway.util.JwtUtil;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cloud.gateway.filter.GatewayFilter;
+import org.springframework.cloud.gateway.filter.factory.AbstractGatewayFilterFactory;
+import org.springframework.http.HttpHeaders;
+import org.springframework.stereotype.Component;
+import org.springframework.web.client.RestTemplate;
+
+@Component
+public class AuthenticationFilter extends AbstractGatewayFilterFactory<AuthenticationFilter.Config> {
+
+    @Autowired
+    private RouteValidator validator;
+
 //    @Autowired
-//    private RouteValidator routerValidator;
-//
-//    @Autowired
-//    private JwtUtil jwtUtil;
-//
-//    @Override
-//    public Mono<Void> filter(ServerWebExchange exchange, GatewayFilterChain chain) {
-//        ServerHttpRequest request = exchange.getRequest();
-//
-//        // Пропуск проверки для открытых API
-//        if (routerValidator.isSecured.test(request)) {
-//            if (this.isAuthMissing(request))
-//                return this.onError(exchange, "Authorization header is missing in request", HttpStatus.UNAUTHORIZED);
-//
-//            final String token = this.getAuthHeader(request).substring(7);
-//
-//            if (jwtUtil.isInvalid(token)) {
-//                System.out.println("Token is invalid: " + token);
-//                return this.onError(exchange, "Authorization header is invalid", HttpStatus.UNAUTHORIZED);
-//            }
-//
-//            // Роль-ориентированная проверка
-//            String role = jwtUtil.getRole(token);
-//            if (!StringUtils.isBlank(role) && !routerValidator.roleBaseApi.check(request, role))
-//                return this.onError(exchange, "Authorization header is invalid", HttpStatus.UNAUTHORIZED);
-//
-//            this.populateRequestWithHeaders(exchange, token);
-//        }
-//        return chain.filter(exchange);
-//    }
-//
-//    private Mono<Void> onError(ServerWebExchange exchange, String err, HttpStatus httpStatus) {
-//        ServerHttpResponse response = exchange.getResponse();
-//        response.setStatusCode(httpStatus);
-//        return response.setComplete();
-//    }
-//
-//    private String getAuthHeader(ServerHttpRequest request) {
-//        return request.getHeaders().getOrEmpty("Authorization").get(0);
-//    }
-//
-//    private boolean isAuthMissing(ServerHttpRequest request) {
-//        return !request.getHeaders().containsKey("Authorization");
-//    }
-//
-//    private void populateRequestWithHeaders(ServerWebExchange exchange, String token) {
-//        Claims claims = jwtUtil.getAllClaimsFromToken(token);
-//        exchange.getRequest().mutate()
-//                .header("id", String.valueOf(claims.get("id")))
-//                .header("role", String.valueOf(claims.get("role")))
-//                .build();
-//    }
-//}
+//    private RestTemplate template;
+
+    @Autowired
+    private JwtUtil jwtUtil;
+
+    public AuthenticationFilter() {
+        super(Config.class);
+    }
+
+    @Override
+    public GatewayFilter apply(Config config) {
+        return ((exchange, chain) -> {
+            if (validator.isSecured.test(exchange.getRequest())) {
+                //header contains token or not
+                if (!exchange.getRequest().getHeaders().containsKey(HttpHeaders.AUTHORIZATION)) {
+                    throw new RuntimeException("missing authorization header");
+                }
+
+                String authHeader = exchange.getRequest().getHeaders().get(HttpHeaders.AUTHORIZATION).get(0);
+                if (authHeader != null && authHeader.startsWith("Bearer ")) {
+                    authHeader = authHeader.substring(7);
+                }
+                try {
+//                    //REST call to AUTH service
+//                    template.getForObject("http://IDENTITY-SERVICE//validate?token" + authHeader, String.class);
+                    jwtUtil.validateToken(authHeader);
+
+                } catch (Exception e) {
+                    System.out.println("invalid access...!");
+                    throw new RuntimeException("un authorized access to application");
+                }
+            }
+            return chain.filter(exchange);
+        });
+    }
+
+    public static class Config {
+
+    }
+}
