+++ gateway/src/main/java/com/fedkoroma/gateway/util/JwtUtil.java
@@ -1,14 +1,13 @@
 package com.fedkoroma.gateway.util;
 
-import io.jsonwebtoken.*;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
-import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import java.security.Key;
-import java.util.Date;
 
 @Component
 public class JwtUtil {
@@ -16,54 +15,10 @@ public class JwtUtil {
     @Value("${JWT_SECRET}")
     public String SECRET;
 
-    private Key key;
-
-    @PostConstruct
-    public void init(){
-        this.key = Keys.hmacShaKeyFor(SECRET.getBytes());
+    public void validateToken(final String token) {
+        Jwts.parserBuilder().setSigningKey(getSignKey()).build().parseClaimsJws(token);
     }
 
-    public Claims getAllClaimsFromToken(String token) {
-        try {
-            return Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token).getBody();
-        } catch (JwtException e) {
-            System.out.println("Ошибка при проверке токена: " + e.getMessage());
-            throw new RuntimeException("Invalid token");
-        }
-    }
-
-    private boolean isTokenExpired(String token) {
-        return this.getAllClaimsFromToken(token).getExpiration().before(new Date());
-    }
-
-    public boolean isInvalid(String token) {
-        System.out.println("Проверяем токен - " + token);
-        return this.validateJwtToken(token) || this.isTokenExpired(token);
-
-    }
-
-    private boolean validateJwtToken(String authToken) {
-        try {
-            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(authToken);
-            return false;
-        }  catch (MalformedJwtException e) {
-            System.out.println("Invalid JWT token: {}" + e.getMessage());
-        } catch (ExpiredJwtException e) {
-            System.out.println("JWT token is expired: {}"+ e.getMessage());
-        } catch (UnsupportedJwtException e) {
-            System.out.println("JWT token is unsupported: {}"+ e.getMessage());
-        } catch (IllegalArgumentException e) {
-            System.out.println("JWT claims string is empty: {}"+ e.getMessage());
-        } catch (Exception e) {
-            System.out.println("-------------------------");
-        }
-
-        return true;
-    }
-
-    public String getRole(String token) {
-        return this.getAllClaimsFromToken(token).get("role").toString();
-    }
 
 
     private Key getSignKey() {
