+++ gateway/src/main/resources/application.properties
@@ -1,26 +1,20 @@
 spring.application.name=gateway
+
 server.port=8765
 
-# ????????? Eureka ???????
 eureka.client.service-url.defaultZone=http://localhost:8761/eureka
-eureka.client.register-with-eureka=true
-eureka.client.fetch-registry=true
 
-
-###############  routing  ###################
-#the gateway enables the function of service registration and discovery,
-# and spring cloud gateway automatically creates a router for each service based on service discovery,
-# and this router will start with the service name.
 spring.cloud.gateway.discovery.locator.enabled=true
-#take service name lower case by default it's uppercase
-spring.cloud.gateway.discovery.locator.lower-case-service-id=true
+spring.cloud.gateway.discovery.locator.lower-case-service-id=true 
 
-gateway.token.validation.type=remote 
-
-spring.cloud.gateway.routes[0].id=client-server
+spring.cloud.gateway.routes[0].id=test
 spring.cloud.gateway.routes[0].uri=lb://client-server
-spring.cloud.gateway.routes[0].predicates[0]=Path=/main/**
+spring.cloud.gateway.routes[0].predicates[0]=Path=/main
+spring.cloud.gateway.routes[0].predicates[1]=Method=GET
+#spring.cloud.gateway.routes[0].filters[2]=AuthenticationFilter
+# ?????? ????????????? ???????????? ??????? AddRequestHeader
+spring.cloud.gateway.routes[0].filters[0]=AuthenticationFilter
 
-spring.cloud.gateway.routes[1].id=security-server
+spring.cloud.gateway.routes[1].id=security-service
 spring.cloud.gateway.routes[1].uri=lb://security-server
 spring.cloud.gateway.routes[1].predicates[0]=Path=/auth/**
